+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Fixed
 - Fixed formatting of `DisplayName` directive.
 - Fixed hover when the symbol below the cursor is the name of a quest which is not found in the workspace.
+- Fixed quest rename.
 
 ## 0.9.0